RaspberryPi-codes/requirements.txt
Adafruit_DHT==1.4.0
RPi.GPIO==0.7.0
spidev==3.5
adafruit-circuitpython-dht==0.0.0
adafruit-blinka==6.19.0
requests==2.26.0 